requirements.txt
numpy
pandas
scipy
torch
pyro-ppl
visdom
matplotlib
seaborn